Sharing Components Between Multiple Pages
how to share one Page component between two routes and show different components on the page based on which route is active.

The goal is to use the Profile component to show the UserProfileCard as well as being able to switch to an editor component when the user wants to update his profile.