Fetching App-Wide Data
we learn how to fetch data that should be available across every page of a Vue.js application. The authenticated user should be available on every page since it's info is displayed in the navigation bar, therefore we can dispatch a call to the Firestore through a Pinia action in the created hook of the root App.vue component.