Preparing the User Interface for Post Updates
We create the user interface for an "Edit Mode" in the PostList component to accommodate editing posts. We also play on a top strength of Vue.js, reusing the PostEditor component for updating posts as well as creating them. Along the way, we also use the post provided to the component to make smart decisions about the button label.