Hiding Content Until Data is Fully Loaded
We improve user experience by removing the incremental showing, hiding all page content until all the data is available. In the Vue.js created hook, we're able to wait for all the data to finish loading and set a ready data property to true that conditionally show/hides the markup in the template.