Show a Loading Spinner while Fetching Async Data
We'll replace the boring loading text with a nice loading spinner. Since we've already implemented the loading functionality for our Vue.js application, adding in the spinner is just a matter of copying some markup from Spinkit, creating a AppSpinner component, and replacing the loading text with the new component.