Signing the User in after Registration
We update the fetchAuthUser action in order to use the id from the logged in user. We also utlize the authStateChange event in order to make sure we set the logged in user to the Pinia state whenever they are logged in. The result? Our application will always know about the user logged in with Firebase Authentication!