Unsubscribing from the Auth User Snapshot[
We tidy up the authenticated user functionality by subscribing to it's snapshot for the entirety of the logged in session and only cleaning up the snapshot after the authentication state has changed. This can be accomplished by providing a callback function to the fetchItems action that is customized for the fetchAuthUser action.